See Dave for details...

+++ Classlib/Persistent.h
@@ -96,7 +96,7 @@ typedef Persistent* PersistentPointer;
 
 inline int Hash(const PersistentPointer& k, int hash_size)
 {
-   return (int)(((long)k^(3*hash_size+1))%hash_size);
+   return (int)((unsigned long)((long)k^(3*hash_size+1))%(unsigned long)hash_size);
 }
 
 inline void Pio(Piostream& stream, char& data) { stream.io(data); }
